refactor(ui): move update node logic from listener to hook
