--- invokeai/frontend/web/src/features/nodes/store/actions.ts
@@ -1,3 +1,0 @@
-import { createAction } from '@reduxjs/toolkit';
-
-export const updateAllNodesRequested = createAction('nodes/updateAllNodesRequested');
